refactor: fix some that fell victim to undo
