+++ harper-core/src/linting/compound_nouns/general_compound_nouns.rs
